Move and refactor connected components into a separate modular file
